+++ Editor/GUI/ContentUpdatePreviewWindow.cs
@@ -10,7 +10,13 @@ namespace UnityEditor.AddressableAssets.GUI
 {
     class ContentUpdatePreviewWindow : EditorWindow
     {
-        internal static bool PrepareForContentUpdate(AddressableAssetSettings settings, string buildPath)
+        /// <summary>
+        /// Opens ContentUpdatePreviewWindow.
+        /// </summary>
+        /// <param name="settings"></param>
+        /// <param name="buildPath"></param>
+        /// <returns>True if successful; false if otherwise (failed to get modified entries).</returns>
+        public static bool PrepareForContentUpdate(AddressableAssetSettings settings, string buildPath)
         {
             var modifiedEntries = ContentUpdateScript.GatherModifiedEntries(settings, buildPath);
             if (modifiedEntries == null)