+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.5] - 2019-07-15
+ - Fixed scenario where scene unload destroys instantiated objects in different scenes. 
+ - Cleaned up SetDirty logic to remove excessive dirtying of assets.
+ 
 ## [1.1.4-preview] - 2019-06-19
  - Fixed an issue where Editor only types were being added to the build.